Extract svg render funcs to new SvgView
