+++ lib/dreadnought_web/live/world_live.html.leex
@@ -1,4 +1,4 @@
-<svg viewbox="-<%= @margin %> -<%= @margin %> <%= @board_size.width %> <%= @board_size.height %>" width="<%= @board_size.width %>" height="<%= @board_size.height %>" class="m-12" overflow="visible" >
+<svg viewbox="-<%= @margin %> -<%= @margin %> <%= @board_size.width %> <%= @board_size.height %>" width="<%= @board_size.width * 2 %>" height="<%= @board_size.height * 2 %>" class="m-12" overflow="visible" >
   <defs>
     <pattern id="play-area-hatch" width="100" height="100" patternUnits="userSpaceOnUse" >
       <rect width="100" height="100" stroke-width="3" stroke="black" fill="blue" />
