+++ lib/dreadnought_web/views/sprite_view.ex
